fix: harden XGo executor lifecycle

+++ tools/xgoexec/capability_js.go
@@ -4,7 +4,7 @@ package xgoexec
 
 import "syscall/js"
 
-func dispatchCapabilityCall(id uint64, name, request string) error {
+func dispatchCapabilityCall(id uint32, name, request string) error {
 	js.Global().Call("xbuilder_xgoexec_capability_call", id, name, request)
 	return nil
 }
