+++ tools/xgoexec/capability_stub.go
@@ -4,6 +4,6 @@ package xgoexec
 
 import "fmt"
 
-func dispatchCapabilityCall(id uint64, name, request string) error {
+func dispatchCapabilityCall(id uint32, name, request string) error {
 	return fmt.Errorf("capability %q is only available in js/wasm", name)
 }
